requirements.txt
requests~=2.26.0
urllib3~=1.24.3
atlassian-python-api~=3.13.0
